refactor: moving logic out of PreferencesRenderingItemFormat.svelte
